move TestBlock to internal/cryptotest
